Removed Play distribution
The Play distribution is no more.  Included in this removal was:

* Removal of skeletons
* Removal of console project
* Removal of all custom SBT setup
* Removal of play and play.bat scripts
* Removal of all SBT tasks related to building the distribution
* Updated the migration guide to document how to obtain Play
