+++ framework/build
@@ -19,4 +19,4 @@ else
 fi
 
 DIR=`dirname "$0"`
-"$JAVA" ${DEBUG_PARAM} -Xms512M -Xmx1536M -Xss2M -XX:ReservedCodeCacheSize=192m -XX:+CMSClassUnloadingEnabled -XX:MaxPermSize=512M ${JAVA_OPTS} -Dfile.encoding=UTF-8 -Dplay.version="${PLAY_VERSION}" -Dplay.home="$DIR" -Dsbt.boot.properties="$DIR/sbt/sbt.boot.properties" ${PLAY_OPTS} -jar "$DIR/sbt/sbt-launch.jar" "$@"
+"$JAVA" ${DEBUG_PARAM} -Xms512M -Xmx1536M -Xss2M -XX:ReservedCodeCacheSize=192m -XX:+CMSClassUnloadingEnabled -XX:MaxPermSize=512M ${JAVA_OPTS} -Dfile.encoding=UTF-8 -Dplay.version="${PLAY_VERSION}" ${PLAY_OPTS} -jar "$DIR/sbt/sbt-launch.jar" "$@"
