--- framework/clean
@@ -1,18 +1,0 @@
-#! /usr/bin/env sh
-
-# Copyright (C) 2009-2013 Typesafe Inc. <http://www.typesafe.com>
-
-echo "[info] Cleaning target directories..."
-
-DIR=`dirname "$0"`
-rm -rf "$DIR/target"
-rm -rf "$DIR/project/project"
-rm -rf "$DIR/project/target"
-rm -rf "$DIR/logs"
-rm -rf "$DIR/src/anorm/target"
-rm -rf "$DIR/src/play/target"
-rm -rf "$DIR/src/templates/target"
-rm -rf "$DIR/src/sbt-plugin/target"
-rm -rf "$DIR/src/console/target"
-rm -rf "$DIR/sbt/boot/scala-2.9.1/play"
-
