--- framework/cleanEverything
@@ -1,11 +1,0 @@
-#! /usr/bin/env sh
-
-# Copyright (C) 2009-2013 Typesafe Inc. <http://www.typesafe.com>
-
-echo "[info] Cleaning everything..."
-
-`dirname $0`/clean
-rm -rf `dirname $0`/../repository/*
-rm -rf `dirname $0`/sbt/boot
-
-echo "[info] Done!"
