--- framework/project/LicenseHelper.scala
@@ -1,11 +1,0 @@
-/*
- * Copyright (C) 2009-2013 Typesafe Inc. <http://www.typesafe.com>
- */
-import sbt._
-
-object LicenseHelper {
-
-  val licenseReport = TaskKey[String]("license-report")
-
-
-}
