--- framework/project/ZipHelper.scala
@@ -1,17 +1,0 @@
-/*
- * Copyright (C) 2009-2013 Typesafe Inc. <http://www.typesafe.com>
- */
-import sbt._
-
-object ZipHelper {
-
-  /**
-   * Uses the native zip command to zip files, because the native command likely supports UNIX permissions.
-   */
-  def zipNative(dir: File, outputZip: File): Unit = {
-    Process(Seq("zip", "-r", outputZip.getAbsolutePath, "."), dir).! match {
-      case 0 => ()
-      case n => sys.error("Failed to run native zip application!")
-    }
-  }
-}
