--- framework/sbt/play.boot.properties
@@ -1,28 +1,0 @@
-#
-# Copyright (C) 2009-2013 Typesafe Inc. <http://www.typesafe.com>
-#
-[scala]
-  version: 2.10.3
-
-[app]
-  org: com.typesafe.play
-  name: console
-  version: 2.3-SNAPSHOT
-  class: play.console.Console
-  cross-versioned: binary
-
-[repositories]
-  local
-  maven-local
-  play-local: file://${play.home}/../repository/local/, [organization]/[module]/[revision]/[type]s/[artifact](-[classifier]).[ext]
-  typesafe-ivy-releases: http://repo.typesafe.com/typesafe/ivy-releases/, [organization]/[module]/[revision]/[type]s/[artifact](-[classifier]).[ext]
-  maven-central
-
-[boot]
-  directory: ${play.home}/sbt/boot
-
-[ivy]
-  ivy-home: ${sbt.ivy.home-${user.home}/.ivy2/}
-  checksums: ${sbt.checksums-sha1,md5}
-  override-build-repos: ${sbt.override.build.repos-false}
-  repository-config: ${sbt.repository.config-${sbt.global.base-${user.home}/.sbt}/repositories}
