--- framework/sbt/sbt.boot.properties
@@ -1,29 +1,0 @@
-#
-# Copyright (C) 2009-2013 Typesafe Inc. <http://www.typesafe.com>
-#
-[scala]
-  version: auto
-
-[app]
-  org: ${sbt.organization-org.scala-sbt}
-  name: sbt
-  version: ${sbt.version-read(sbt.version)[0.13.0]}
-  class: ${sbt.main.class-sbt.xMain}
-  components: xsbti,extra
-  cross-versioned: ${sbt.cross.versioned-false}
-  
-[repositories]
-  local
-  maven-local
-  play-local: file://${play.home}/../repository/local/, [organization]/[module]/[revision]/[type]s/[artifact](-[classifier]).[ext]
-  typesafe-ivy-releases: http://repo.typesafe.com/typesafe/ivy-releases/, [organization]/[module]/[revision]/[type]s/[artifact](-[classifier]).[ext], bootOnly
-  maven-central
-
-[boot]
-  directory: ${play.home}/sbt/boot
-
-[ivy]
-  ivy-home: ${sbt.ivy.home-${user.home}/.ivy2/}
-  checksums: ${sbt.checksums-sha1,md5}
-  override-build-repos: ${sbt.override.build.repos-false}
-  repository-config: ${sbt.repository.config-${sbt.global.base-${user.home}/.sbt}/repositories}
