--- framework/skeletons/java-skel/app/controllers/Application.java
@@ -1,14 +1,0 @@
-package controllers;
-
-import play.*;
-import play.mvc.*;
-
-import views.html.*;
-
-public class Application extends Controller {
-
-    public static Result index() {
-        return ok(index.render("Your new application is ready."));
-    }
-
-}
