--- framework/skeletons/java-skel/app/views/index.scala.html
@@ -1,7 +1,0 @@
-@(message: String)
-
-@main("Welcome to Play") {
-
-    @play20.welcome(message, style = "Java")
-
-}
