--- framework/skeletons/java-skel/build.sbt
@@ -1,12 +1,0 @@
-name := "%APPLICATION_NAME%"
-
-version := "1.0-SNAPSHOT"
-
-libraryDependencies ++= Seq(
-  javaJdbc,
-  javaEbean,
-  javaWs,
-  cache
-)
-
-play.Project.playJavaSettings
