--- framework/skeletons/scala-skel/README
@@ -1,4 +1,0 @@
-This is your new Play application
-=====================================
-
-This file will be packaged with your application, when using `play dist`.
