--- framework/skeletons/scala-skel/app/controllers/Application.scala
@@ -1,12 +1,0 @@
-package controllers
-
-import play.api._
-import play.api.mvc._
-
-object Application extends Controller {
-
-  def index = Action {
-    Ok(views.html.index("Your new application is ready."))
-  }
-
-}
