--- framework/skeletons/scala-skel/project/build.properties
@@ -1,1 +1,0 @@
-sbt.version=%SBT_VERSION%
