--- play
@@ -1,57 +1,0 @@
-#! /usr/bin/env sh
-
-PRG="$0"
-# resolve relative/absolute symlinks
-while [ -h "$PRG" ] ; do
-  ls=`ls -ld "$PRG"`
-  link=`expr "$ls" : '.*-> \(.*\)$'`
-  if expr "$link" : '/.*' > /dev/null; then
-    PRG="$link"
-  else
-    dir=`dirname "$PRG"`
-    PRG="$dir/$link"
-  fi
-done
-dir=`dirname "$PRG"`
-
-if [ -z "$JAVA_HOME" ]; then
-  JAVA="java"
-else
-  JAVA="$JAVA_HOME/bin/java"
-fi
-
-if [ -f conf/application.conf -o -f conf/reference.conf ] || [ -d project ]; then
-  if test "$1" = "stop"; then
-    if [ -f RUNNING_PID ]; then
-      echo "[info] Stopping application (with PID `cat RUNNING_PID`)..."
-      kill `cat RUNNING_PID`
-
-      RESULT=$?
-
-      if test "$RESULT" = 0; then
-        echo "[info] Done!"
-        exit 0
-      else
-        echo "[\033[31merror\033[0m] Failed ($RESULT)"
-        exit $RESULT
-      fi
-    else
-      echo "[\033[31merror\033[0m] No RUNNING_PID file. Is this application running?"
-      exit 1
-    fi
-  fi
-  
-  if test "$1" = "debug"; then
-    JPDA_PORT="9999"
-    shift      
-  fi
-
-  if [ -n "$1" ]; then
-    JPDA_PORT="${JPDA_PORT}" "$dir/framework/build" "$@"
-  else
-    JPDA_PORT="${JPDA_PORT}" "$dir/framework/build" play
-  fi
-  
-else
-  "$JAVA" -Dplay.home="$dir/framework" -Dsbt.boot.properties="$dir/framework/sbt/play.boot.properties" ${PLAY_OPTS} -jar "$dir/framework/sbt/sbt-launch.jar" "$@"
-fi
